+++ GooglePay/src/test/java/com/braintreepayments/api/GooglePayInternalClientUnitTest.kt
@@ -19,7 +19,6 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import java.util.concurrent.CountDownLatch
 
-
 @RunWith(RobolectricTestRunner::class)
 class GooglePayInternalClientUnitTest {
 
@@ -81,7 +80,6 @@ class GooglePayInternalClientUnitTest {
             assertNull(error)
             countDownLatch.countDown()
         }
-
         countDownLatch.await()
     }