+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * GooglePay
     * Deprecate `googleMerchantId`
 * SharedUtils
-  * Use char array to hold `HttpRequest` data. Dispose data immediately after making http request.
+  * Use byte array to hold `HttpRequest` data. Dispose data immediately after making http request.
 
 ## 4.8.0